Init complete: CLI functionality, requirements and updated README (#2)
* Revise LICENSE with new copyright and conditions

Updated copyright holder and added usage restrictions.

* CLI setup for basic_trimmer

Configures the env to use setuptools for building.

This includes adding a build-system section to pyproject.toml, specifying the package directory, and defining an entry point for the basic trimmer script.
It also marks the source as editable in the lock file.

* Updates README with CLI command

* Added "requirements.txt"

Alternative to using pyproject.toml.